+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - SDK version added to election metadata.
+- Added new function `changeElectionCensus` to change the election census dynamically.
 
 ### Changed